idiomatic prefix declaration, restructure tests
