+++ hdl21/external_module.py
@@ -15,6 +15,7 @@ from .params import HasNoParams, isparamclass, _unique_name
 from .signal import Signal, Visibility
 from .instance import calls_instantiate
 from .qualname import qualname_magic_methods
+from .props import Properties
 
 @dataclass
 @qualname_magic_methods
@@ -38,7 +39,7 @@ class ExternalModule:
     paramtype: Type = HasNoParams  # Parameter-type `paramclass`
     desc: Optional[str] = None  # Description
     domain: Optional[str] = None  # Domain name, for references upon export
-    devicetype: Optional[str] = None # Device type, takes its cue from primitives.py
+    props: Properties = Properties() # Additional properties, for use by the user.
 
     @property
     def ports(self) -> dict:
